Update README.md
Following the instructions as they were before this commit would install "pcgr" in the base environment, which is not good practice. Note also that conda build does not create a dedicated environment which can be activated, meaning that "conda activate pcgr" would not work. The now modified instructions address this issue.