Add files via upload
Added new version of the archive. The python script "tableize_vcf.py" has been updated from python2 to python3 syntax. Note that the versioning of the archive should be updated to reflect for this change, but I am a bit unsure how to proceed given this is third party software.